Python3 compatibility fixes
sudo python3 measure.py
Traceback (most recent call last):
   File "measure.py", line 3, in <module>
     bmp = bmp183()
   File "/home/pi/Programmation/Python/bmp183/bmp183.py", line 86, in
__init__
     self.measure_pressure()
   File "/home/pi/Programmation/Python/bmp183/bmp183.py", line 201, in
measure_pressure
     self.calculate_pressure()
   File "/home/pi/Programmation/Python/bmp183/bmp183.py", line 210, in
calculate_pressure
     self.B3 = (((self.AC1 * 4 + X3) << self.BMP183_CMD['OVERSAMPLE_3'])
+ 2 ) / 4
TypeError: unsupported operand type(s) for <<: 'numpy.float64' and 'int'

Suggested by Thomas Legrand

+++ bmp183.py
@@ -206,7 +206,7 @@ class bmp183():
 		X1 = (self.B2 * (self.B6 * self.B6 / 2**12)) / 2**11
 		X2 = self.AC2 * self.B6 / 2**11
 		X3 = X1 + X2
-		self.B3 = (((self.AC1 * 4 + X3) << self.BMP183_CMD['OVERSAMPLE_3']) + 2 ) / 4
+		self.B3 = (((numpy.uint32(self.AC1 * 4 + X3)) << self.BMP183_CMD['OVERSAMPLE_3']) + 2 ) / 4
 		X1 = self.AC3 * self.B6 / 2**13
 		X2 = (self.B1 * (self.B6 * self.B6 / 2**12)) / 2**16
 		X3 = ((X1 + X2) + 2) / 2**2
